+++ 3_scripts/recordcleaner.py
@@ -176,9 +176,9 @@ if __name__ == "__main__":
         logging.info('\n#> A4: HUH name query\n')
         
         # print('This takes a moment to initialise...')
-        # tmp_occs_3 = huh_query.huh_wrapper(tmp_occs_3, verbose = True, debugging = False)
-        # tmp_occs_3 = tmp_occs_3.reset_index(drop=True)
-        # logging.info('\n #> STEP A complete.\n')
+        #tmp_occs_3 = huh_query.huh_wrapper(tmp_occs_3, verbose = True, debugging = False)
+        tmp_occs_3 = tmp_occs_3.reset_index(drop=True)
+        logging.info('\n #> STEP A complete.\n')
 
 
         ###------------------------------------------ Step B -------------------------------------------####
@@ -199,7 +199,6 @@ if __name__ == "__main__":
 
         print(tmp_occs_4[['recorded_by','colnum','ddlat']])
 
-        stop
         
         # Double checking duplication stats, should show 0. (i.e. repeat B1)
         logging.info('\n#> B2: Duplicates - stats after first merge \n')